Add router with DI method
