+++ GoogleMapsTestApp/Protocols/HasRouter.swift
@@ -1,0 +1,13 @@
+//
+//  HasRouter.swift
+//  GoogleMapsTestApp
+//
+//  Created by Andrey on 06.11.2019.
+//  Copyright © 2019 Andrey. All rights reserved.
+//
+
+import Foundation
+
+protocol HasRouter: class {
+    var router: Router! { get set }
+}
