+++ GoogleMapsTestApp/Protocols/Router.swift
@@ -1,0 +1,13 @@
+//
+//  Router.swift
+//  GoogleMapsTestApp
+//
+//  Created by Andrey on 06.11.2019.
+//  Copyright © 2019 Andrey. All rights reserved.
+//
+
+import Foundation
+
+protocol Router: class {
+    func route(from controller: UIViewController)
+}
