+++ GoogleMapsTestApp/Struct/Constants.swift
@@ -1,0 +1,14 @@
+//
+//  Constants.swift
+//  GoogleMapsTestApp
+//
+//  Created by Andrey on 06.11.2019.
+//  Copyright © 2019 Andrey. All rights reserved.
+//
+
+import Foundation
+
+struct Constants {
+    static let buckets: [NSNumber] = [4, 5, 6, 8, 10]
+    static let colors: [UIColor] = [.red, .blue, .cyan, .gray, .green]
+}
